+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - 2023-11-05
+## [0.1.0.0] - 2023-11-09
+
+* `baseLength` and `decisionLength` functions
+* rewritten the counting functions
+
+## [0.0.1.0] - 2023-11-05
 
 Initial version